Mistaken approach to suggestions
I started writing logic for a world where "suggestions" is a staging
table that attempts to mirror the structure of "topics", and not one
where each recieved suggestion is its own document (the way it should
be). I realized why this would be a bad idea a few minutes later, so I'm
committing now and backing up.

To explain my thinking as to why this is a bad idea: Each suggestion
really has its own point-of-view that should be considered individually,
rather than immediately lumping it together with all other suggestions.
You have to deal with multiple path suggestions for one site (eg. "Darth
Vader" being suggested as both "Anakin Skywalker" and "Darth Vader") and
all kinds of other differences that shouldn't be conflated into one
unvetted entry.