Working-as-intended (and not just running) index.js
Changed the API method routes to just be static paths, with all the
params being form/query ones. After much deliberation, this is what I
decided when comparing this approach to the prior one where the topic
name was part of the route:

Benefits of topic name being route parameter:
  - Adheres to theoretical benefit of having hierarchical paths that
    never really actually matters in practice

Benefits of topic name being query parameter:
  - Fixed URLs don't have to be constructed (via script, etc.) (meaning
    you can get/post with form data, possibly even skipping an XHR
    entirely)
  - Makes it so topic parameter isn't mandatory (for instance, you could
    get the pagelist and topic name based on only having one of the
    URLs)
  - Makes the whole '/topic/' prefix kind of obsolete, actually

With this commit I'm also officially embracing the name of this project
as "metapoint" rather than "wikilinker". Cheers!